[AIP-21] Move contrib.hooks.azure_cosmos_hook airflow.contrib.hooks.azure_cosmos_hook
