+++ tests/providers/microsoft/azure/hooks/test_azure_cosmos.py
@@ -26,9 +26,9 @@ import uuid
 
 import mock
 
-from airflow.contrib.hooks.azure_cosmos_hook import AzureCosmosDBHook
 from airflow.exceptions import AirflowException
 from airflow.models import Connection
+from airflow.providers.microsoft.azure.hooks.azure_cosmos import AzureCosmosDBHook
 from airflow.utils import db
